split the cart page into multiple components
